Lab-37/requirements.txt
# Strava MCP Server Dependencies (Python 3.8 compatible)
# HTTP and API communication
requests==2.31.0          # HTTP library for making API calls to Strava

# Configuration and environment management
python-dotenv==1.0.0      # Load environment variables from .env files

# AI and machine learning
openai==1.3.7             # OpenAI API client for Azure OpenAI integration
